refactor(ds): abstract out logics
